Moved python files to src/id3_variants_training folder.
