+++ src/id3_variants_training/ConfusionMatrix.py
@@ -1,7 +1,6 @@
-from ID3_Class import ID3
-from local_API import LOCAL_API
-from ga4gh_API import GA4GH_API
-from ID3_Node import ID3_Node
+from src.id3_variants_training.ID3_Class import ID3
+from src.id3_variants_training.local_API import LOCAL_API
+from src.id3_variants_training.ID3_Node import ID3_Node
 
 class ConfusionMatrix(ID3):
