+++ src/id3_variants_training/main.py
@@ -1,4 +1,4 @@
-from ID3_Class import ID3
+from src.id3_variants_training.ID3_Class import ID3
 import argparse
 
 parser = argparse.ArgumentParser()
